refactor: rename loader -> data
